requirements.txt
# G1 Inspire Pipeline - Base Requirements
# 각 공식 저장소별 추가 의존성은 해당 저장소 참조

# Core
numpy>=1.24.0
pyyaml>=6.0
torch>=2.0.0

# Simulation
mujoco>=3.0.0

# Visualization
pillow>=9.0.0
matplotlib>=3.5.0

# Note: 추가 의존성
# - LeRobot: pip install -e 'external/lerobot[unitree_g1]'
# - Isaac Gym: NVIDIA 공식 설치 가이드 참조
# - Unitree SDK2: CycloneDDS 필요 (https://github.com/unitreerobotics/unitree_sdk2_python)
